Renaming dclist.rs and list.rs.
