+++ hfo2/src/lib.rs
@@ -35,13 +35,13 @@ mod dlog;
 mod api;
 mod arch;
 mod cpu;
-mod dclist;
 mod list;
 mod memiter;
 mod mm;
 mod mpool;
 mod page;
 mod panic;
+mod slist;
 mod spci;
 mod spinlock;
 mod std;
